Feature/ws game session

+++ client/src/core/components/issueCard/IssueCard.tsx
@@ -6,8 +6,6 @@ import { Text } from 'core/components/Text';
 import { useTypeSelector } from 'core/hooks/useTypeSelector';
 import { Roles } from 'core/types/roleType';
 import { deleteIssueById } from 'core/api/issues.service';
-import { useDispatch } from 'react-redux';
-import { getIssues } from 'store/actionCreators/issue';
 import styles from './IssueCard.module.scss';
 interface IssueCardProp {
   title: string;
@@ -17,6 +15,7 @@ interface IssueCardProp {
   modalShowEdit?: (flag: boolean) => void;
   handleIssueId: (id: string) => void;
   gameMode?: boolean;
+  resultMode?: boolean;
   currentIssueId?: string;
 }
 
@@ -28,47 +27,49 @@ export const IssueCard: React.FC<IssueCardProp> = ({
   modalShowEdit,
   handleIssueId,
   gameMode,
+  resultMode,
   currentIssueId,
 }) => {
   const { currentUser } = useTypeSelector(state => state.currentUser);
-  const dispatch = useDispatch();
   const className = currentIssueId !== id ? `${styles.issue}` : `${styles.issue} ${styles.green}`;
 
+  const handleEditIssue = () => {
+    if (modalShowEdit) {
+      modalShowEdit(true);
+    }
+    handleIssueId(id);
+  };
+
+  const handleDeleteIssue = () => {
+    if (modalShowDelete) {
+      modalShowDelete(true);
+    }
+    handleIssueId(id);
+  };
+
+  const issueBtns = (): JSX.Element => {
+    if (!resultMode && gameMode && currentUser.role === Roles.creator) {
+      return <CloseIcon className={styles.btn} onClick={() => deleteIssueById(id)} />;
+    } else if (!resultMode && !gameMode) {
+      return (
+        <>
+          <EditIcon className={styles.btn} onClick={handleEditIssue} />
+          <DeleteOutlineIcon className={styles.btn} onClick={handleDeleteIssue} />
+        </>
+      );
+    }
+    return <></>;
+  };
+
   return (
-    <div className={styles.issue}>
+    <div className={className}>
       <div className={styles.title}>
         <Text textLvl={'base'}>{title}</Text>
         <div className={styles.priority}>
           <Text textLvl={'comment'}>{priority} priority</Text>
         </div>
       </div>
-      <div>
-        {gameMode && currentUser.role === Roles.creator ? (
-          <CloseIcon className={styles.btn} onClick={() => deleteIssueById(id)} />
-        ) : null}
-        {!gameMode ? (
-          <>
-            <EditIcon
-              className={styles.btn}
-              onClick={() => {
-                if (modalShowEdit) {
-                  modalShowEdit(true);
-                }
-                handleIssueId(id);
-              }}
-            />
-            <DeleteOutlineIcon
-              className={styles.btn}
-              onClick={() => {
-                if (modalShowDelete) {
-                  modalShowDelete(true);
-                }
-                handleIssueId(id);
-              }}
-            />
-          </>
-        ) : null}
-      </div>
+      <div>{issueBtns()}</div>
     </div>
   );
 };
